- Utilities and bug fixes.
git-svn-id: svn://localhost/tmp/repo1/trunk@390 09e1528e-719b-4268-a937-68dd4ef2d4b4

+++ OSN/main/src/it/unitn/disi/newscasting/experiments/schedulers/distributed/MasterImpl.java
@@ -53,7 +53,7 @@ public class MasterImpl implements IMaster, IMasterAdmin, Runnable {
 		synchronized (fExperiments) {
 			for (int i = 0; i < fExperiments.length; i++) {
 				fExperiments[i] = new ExperimentEntry(
-						iterator.nextIfAvailable());
+						(Integer) iterator.nextIfAvailable());
 			}
 			Arrays.sort(fExperiments);
 		}
@@ -194,7 +194,8 @@ public class MasterImpl implements IMaster, IMasterAdmin, Runnable {
 			throws RemoteException {
 		WorkerEntry entry = fWorkers.get(workerId);
 		if (entry == null) {
-			throw new InvalidWorkerException("Worker " + workerId + " is not valid.");
+			throw new InvalidWorkerException("Worker " + workerId
+					+ " is not valid.");
 		}
 
 		for (int i = 0; i < fExperiments.length; i++) {
